requirements.txt
alembic==1.0.11
amqp==2.5.1
apache-airflow==1.10.6
apispec==2.0.2
argcomplete==1.10.0
astroid==2.3.3
attrs==19.1.0
Babel==2.7.0
billiard==3.6.1.0
boto3==1.10.14
botocore==1.13.14
cached-property==1.5.1
celery==4.3.0
certifi==2019.6.16
chardet==3.0.4
click==6.7
colorama==0.4.1
colorlog==4.0.2
configparser==3.5.3
croniter==0.3.30
defusedxml==0.6.0
dill==0.2.9
docopt==0.6.2
docutils==0.15.2
dumb-init==1.2.2
feedparser==5.2.1
Flask==1.1.1
Flask-Admin==1.5.3
Flask-AppBuilder==1.13.1
Flask-Babel==0.12.2
Flask-Caching==1.3.3
Flask-JWT-Extended==3.21.0
Flask-Login==0.4.1
Flask-OpenID==1.2.5
Flask-SQLAlchemy==2.4.0
flask-swagger==0.2.13
Flask-WTF==0.14.2
flower==0.9.3
funcsigs==1.0.0
future==0.16.0
gitsome==0.8.4
graphviz==0.13.2
gunicorn==19.9.0
idna==2.8
importlib-metadata==0.19
iso8601==0.1.12
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.1
jmespath==0.9.4
json-merge-patch==0.2
jsonschema==3.0.2
kombu==4.6.4
lazy-object-proxy==1.4.2
lockfile==0.12.2
Mako==1.1.0
Markdown==2.6.11
MarkupSafe==1.1.1
marshmallow==2.19.5
marshmallow-enum==1.5.1
marshmallow-sqlalchemy==0.17.0
mccabe==0.6.1
more-itertools==7.2.0
mysqlclient==1.4.4
numpy==1.17.0
ordereddict==1.1
pandas==0.25.1
pendulum==1.4.4
pipenv==2018.11.26
ply==3.11
prison==0.1.0
prompt-toolkit==2.0.9
psutil==5.6.3
Pygments==2.4.2
PyJWT==1.7.1
pylint==2.4.3
pyrsistent==0.15.4
python-daemon==2.1.2
python-dateutil==2.8.0
python-editor==1.0.4
python3-openid==3.1.0
pytz==2019.3
pytzdata==2019.2
PyYAML==5.1.2
requests==2.22.0
s3transfer==0.2.1
setproctitle==1.1.10
six==1.12.0
SQLAlchemy==1.3.7
tabulate==0.8.3
tenacity==4.12.0
termcolor==1.1.0
text-unidecode==1.2
thrift==0.11.0
tornado==5.1.1
typed-ast==1.4.0
tzlocal==1.5.1
unicodecsv==0.14.1
uritemplate==3.0.0
uritemplate.py==3.0.2
urllib3==1.25.3
vine==1.3.0
virtualenv==16.7.7
virtualenv-clone==0.5.3
wcwidth==0.1.7
Werkzeug==0.15.5
wrapt==1.11.2
WTForms==2.2.1
zipp==0.6.0
zope.deprecation==4.4.0